Incorporate sprint review feedback.
